refacto chat components.
