+++ front/src/components/chat/chat-create/FormHeader.tsx
@@ -1,6 +1,6 @@
 import { SetStateAction } from 'react';
-import '../../../../style/components/chat/chat-container/create-channel-form/form-header.css';
-import Avatar from '../../../Avatar';
+import '../../../style/components/chat/chat-container/create-channel-form/form-header.css';
+import Avatar from '../../Avatar';
 
 const FormHeader = ({
   avatar,
