+++ front/src/components/chat/components/components/ChannelCard.tsx
