+++ front/src/pages/chat/chat-channel/ChannelMessaging.tsx
@@ -1,8 +1,8 @@
 import { useUser } from '../../../hooks/useUser';
 import { User } from '../../../types/User.type';
-import ChatBody from '../../../components/Chat/ChatContainer/ChannelMessaging/ChatBody';
-import JoinChannel from '../../../components/Chat/ChatContainer/ChannelMessaging/JoinChannel';
 import useChannel from '../../../hooks/useChannel';
+import JoinChannel from '../../../components/chat/chat-channel/channel-messaging/JoinChannel';
+import ChatBody from '../../../components/chat/chat-channel/channel-messaging/ChatBody';
 
 export default function ChannelMessaging() {
   const { myId } = useUser();
